Add new entry to admin cookbook and change title (#144)
Restoring power to the station is a useful and cook-book-y action. It's
definitely something that should be documented in the wiki somewhere.
Even though it isn't a command, no other place seems better for it than
the admin command cookbook. This commit changes the cookbook's title to
encompass admin cook-book-y things in general, and it adds the new entry
to it.