Neutral Antag proposal: Pizza Delivery Critter (#195)
# Arnold's Pizza Needs You!


![image](https://github.com/space-wizards/docs/assets/732532/51cb9f9e-9e9c-41fe-abff-79edd8042e6e)

- Alone?
- Gullible?
- Poor?
- Unaware of our reputation?

If most/all of these apply to you, _or_ we've successfully kidnapped you
from your home, then Arnolds Pizza has a job with your name on it!

We need delivery critters willing to do one of the most difficult and
thankless tasks in known space: delivering pizzas to NanoTrasen
employees!

NanoTrasen knows we make the best pizza in the galaxy*. That's why they
have an exclusive contract with us to fulfil their Random Reward Pizza
employee satisfaciton scheme. However, NanoTrasen's boutique, black-ops
and (frequently) badly-maintained space stations are both challenging to
access and deliver to. The fell-off-a-back-of-a-van BlueSpace
teleporation devices that we use for such challenging customers are both
extremely unethical *and\* have a carry-limit small enough that we can't
send a full-sized sapient creature through with our KeepWarm pizza boxes
in tow!

That means there's a _unique_ and _exciting_ opportunity for small,
light, _expendable_ critters in our delivery services. Join today, and
we promise that we'll let you leave afterward. Just remember to get that
receipt!

*As do the Syndicate, but we don't tell them that.

---

Note that this PR makes no suggestion for what the delivery critters
should look like. The illustrative art above is just where my brainrot
took me when designing this feature.