Game-Area documentation overhaul, Work Groups and Maintainer/Review Policy (#199)
The biggest change here is separating game (SS14) documentation into
Game Areas to better categorize mechanics/docs and make it easier to
find things.
Game Areas are separated into Global and Department areas:
- A Global Area is a part of the game that does not fit under a single
department or is specific to the individual player's experience. These
may also be high-level concepts related to the game. (For example: UI,
Art, Mapping, Species, etc.)
- A Department Area is a part of the game that fits the design/theme of
a specific in-game department. (For Example: Security, Atmospherics,
Command, etc.)

This PR also implements workgroups as a concept (Code-Owners but for
Game Design). A workgroup is an opt-in/opt-out role that is responsible
for maintaining the documentation, design, and PR Guidelines of a Game
Area. Maintainers can be a part of multiple workgroups or none if they
so choose.

Additionally, this PR will define policy for PR reviews, specifically
going over the review process and what is required to approve/merge a
PR. Along with a list of guidelines for contributors to look through to
help speed up the review process. This also includes a Maintainer policy
document that outlines the powers/expectations that the maintainer role
entails, how maintainers are chosen, and what contributors should do if
they want to become a maintainer someday.