Reduced Metagaming Mechanics (#214)
This proposal aims to reduce the impact metaknowledge has on the game by
shifting it from being a rules-enforced system to a mechanics-enforced
system. This will reduce ambiguity for what constitutes a rulebreak,
reduce admin workload and better align player knowledge with character
knowledge.

The proposal identifies 5 "problems" currently in the game where
knowledge about the game is restricted via rules enforcement, and
suggests changes that allow players to possess and act upon the
knowledge while making it harder to do so without proper justification.

The important part of the proposal is the identification of issues and
attempts to solve them; the particulars of a specific solution (e.g. the
damage proposed in the "Implant Problem") is malleable based on
feedback.

---------

Co-authored-by: Kara <lunarautomaton6@gmail.com>