Changeling antagonist (#263)
Everyone knows the changeling, the space monster that eats people and
then turns into them.
Based mainly on goonstation's changelings with a few elements from tg,
just the armblade and stinger chemicals.

The main thing about changeling is that it round-removes it's victims
rather than temporarily subduing them. Everyone is a victim, not just a
select few people. To make up for this, the changeling has the ability
to create ghost roles that assist it in it's goals. Creating ghost roles
is optional for the ling, it is not obligated to bring it's victims back
into the round.
The distinguishing feature about this version of the changeling is that
it's humanoid form is it's unnatural form, it actually looks like a
horrific flesh monster (depicted below), and it can choose to reveal
it's true form as a combat mode of sorts to grant itself increased
survivability and better offensive abilities.

![pixil-gif-drawing - 2024-08-07T055243
323](https://github.com/user-attachments/assets/0f5a951d-353b-4c32-8efa-2cf2664d99c1)

Feedback is greatly appreciated.