Add toolshed examples (#277)
This PR includes some minor alterations to the introduction to toolshead
and adds a ton of examples for how to use toolshead. The examples are all
from "Milon's Guide to Toolshed" which has been privately shared for some
time. The content was published with permission from the author. It was originally
targeted at game admins, however should also be useful to developers.

---------

Co-authored-by: Milon <plmilonpl@gmail.com>
Co-authored-by: 0x6273 <0x40@keemail.me>