Added Exemptions section to review-procedure.md (#366)
The main content of this PR is adding an exemption for "Maintainer
Maptainers" in regards to mapping PRs. I've also taken the opportunity
to clean up the document by moving any other exemption into the new
"Exemptions" section, instead of having them spread out in different
places; these have not been changed, but the language has been updated
for clarity.