Update outdated/unclear parts of the mapping guide (#358)
After many headaches trying to follow the mapping guide as written and
not understanding why it's not working properly, I am humbly offering
these updates to make it easier for people like me who are just starting
on SS14 mapping to understand what's going on:

- Map ID needs to be relatively high to avoid colliding with existing
IDs. I've heard that most people just straight up use 1000, so updating
all the IDs to that to establish best practices (+ explainer as to why)
- loadmapacts is deprecated, it is now mappingclientsidesetup
- Trying to be a bit more clear how to force your map and start it both
on Tools and Release build
- Adding the power command to the quick setup, since it's extremely
useful in most cases
- Explain that saved maps go into the Content.Server folder more
clearly, and you have to manually move them to Resources/Maps later.
- Explain that the Content.Server saved map has to be manually deleted
later on, or it will persist.
- Very slightly improve the explanation of the map image tool to make it
clear where you need to run it, especially if you don't use VS.
- Update the pipe color hex codes to the ones actually used on the live
version. Those "muted" pipe colors have not been used for a long time
and it's impossible to find the hex codes otherwise.