Design Document: Security Department (#399)
This PR includes a rewrite of the Security Department design document,
following the template from #397. It's originally based on a proposal by
@keronshb.

This is a directional design document, meaning the goal is to define
boundaries and goals for the game area (in this case, the Security
department). It is not aiming to provide specific mechanics or
implementations, and instead act as a guidance for future PRs in this
area, both for contributors when making PRs and Maintainers when
evaluating them.

Should this PR be merged it will act as a reference point for future
departmental and game area design docs, as an example for how a
directional documents should look.