Add implementation plan requirement to design docs (#435)
Discussed internally in
https://forum.spacestation14.com/t/require-implementation-plan-in-design-doc-proposals/18637
To summarize, a lot of design docs lack critical details about how the
technical implementation is planned, and some proposed features were
literally impossible in our game engine. Other docs had been accepted a
long time ago, but have never been implemented.
Both of these may happen if the author of the doc does not have the
required coding knowledge themselves. To solve this we will require
authors of new design docs to confirm that they plan on implementing the
proposed feature either themselves, or have someone else do it for them.
This helps maintainers to focus on design docs that actually have a
chance to be implemented in the foreseeable future.
Existing docs will not be affected.