coding conventions update (#418)
Updates the coding conventions with a bunch of new info. Summary:
- Added a section on TimeSpans and pausing entities. It's short and
simple enough that I don't think it deserves a separate page, and I
think I prefer having it all in one place.
- Added a section on Field Deltas. Sloth should review it because I
pulled these numbers out of my ass based on my poor understanding of the
code.
- Standarized docstring to be indented.
- Standarized JSON to use 4 space indentation and added fix as you go,
if you're going to add new states. The reasoning behind it is that Rider
will force you to use 4 spaces because of editorconfig, and it's a
massive PITA to disable it just because some file uses 2 spaces, then
re-enable it every time. We should just stick to 4 space if we're
touching a file.
- Using SoundCollections over SoundPaths whenever possible.
- Other minor style changes.

---------

Co-authored-by: slarticodefast <161409025+slarticodefast@users.noreply.github.com>