[Proposal] Playtime Reminders - A feature raising awareness of addiction to SS14 (#439)
Title. This is a 9154 character proposal document for a feature that can
be explained in a single sentence.

This is a proposal to add a feature that tells players how much time
they've spent on the game in a given day.

Within the proposal, we've outlined our motivation in implementing this
feature. A major contributor to toxicity across the SS13 and SS14
community is addiction to the game. There is no silver bullet in solving
this. But frankly, as a game designer, we feel that it's a moral
obligation to help curb it by raising awareness.

The PR associated with this proposal is
space-wizards/space-station-14#36483.