Roundstart Species Design Guidelines (#477)
# Species Guidelines

This document covers the policy for the acceptance and maintenance of
proposed playable roundstart species on Wizard's Den servers. This
document exists as both the direct rules and a further set of soft
principles for proposing, developing and maintaining roundstart species
for Space Station 14.

---

I've written this as I've threatened to a few times on the Discord and
finally decided to do so this evening.

There's probably more than one alternative proposal in the works for
this; this is ergo something of a focus point of discussion as well as a
proposal in its own rights.

I've drafted these rules and guidelines after a lot of discussion on the
Wizden Discord, on and off, pretty much since the start of 2024. They
encapsulate my best effort at covering the views I encountered in these
discussions, which I have then synthesized and, inevitably, coloured
with my own opinion and views.

I'm happy to iterate on this document in response to feedback.

---------

Co-authored-by: SlamBamActionman <83650252+SlamBamActionman@users.noreply.github.com>
Co-authored-by: ScarKy0 <106310278+ScarKy0@users.noreply.github.com>