Tidy up Triage Proceedure document (#493)
This PR tidies the triage document, reducing repeated information and
trying to cover a few esoteric things, like the role of some
Maintainer-only tags.

There's one thing I need feedback on the wording of here - one important
thing about P0s is that the ultimate sign-off on if something is P0 or
not is down to the project leader, but that's not really official
policy, just sort of the nature of what P0s signify. I can not include
any mention of it. but it feels kinda notable as a point about how P0s
work from the perspective of someone triaging.

Maints: feel free to edit this doc directly, I can handle the Git on my
end if I want to change any wording.