Atmospherics Design Document (#492)
This PR fleshes out the Atmospherics design document, laying the
groundwork for what Atmospherics is as a department and how players
interact with it and its systems. This ddoc will be used as a foundation
to write more design documents delving into the various mechanics,
future systems, and what things are supposed to be like in Atmospherics.

The merging of this document will grant powers to the Atmospherics side
of the Engineering/Atmospherics Workgroup.