Combat Design Document (#502)
This is a game area design document to outline the design of SS14's
combat.

Note that it is meant to be a general overview of what we're aiming for
with SS14 combat and therefore doesn't dive into specifics such as
weapon types, stats or specific roles. Basically acts as a base that the
Combat workgroup can expand upon with further design docs.

Credit to Keron, Southbridge and AdmiralOblivious for contributing with
the initial version of the document.