Medical design doc(s) (#510)
I have made two documents in related to the (currently un)official
medical workgroup.

The documents comprise of a general overview of what medical gameplay
should look like from an overarching design standpoint, and a second
which goes over the goals and responsibilities of the medical workgroup.

I saw it as important to include the latter since due to the current and
unique nature of medical development, we need a bit more freedom in
terms of managing, merging, freezing, and closing medical related PRs
than a design doc would offer. It also clearly outlines the scope of
what we're doing so if people have questions they can get answers, and
if people want to contribute they can know what they can contribute to
and how.

---------

Co-authored-by: opl- <opl-@users.noreply.github.com>