Traitors Design Document (Real) (#554)
## Description
Exactly as the title says, this is a design document for the traitors
antagonist.
You already know why this exists if you're staff. Traitors is the most
common gamemode and is the number one source of a lot of game design
pains and everyone has different ideas on what needs to be done.

Collectively though, it's agreed something does need to be done and
without a proper design and plan written down, Traitors as an antagonist
and gamemode is just going to suffer.

This is a mald PR btw, I made this cause I was tired of constant talk
about what traitors need and no action being done (Yes, I know I'm a
hypocrite considering I also actively engage in these discussion and am
easily top 3 yappers on the topic).

If you have a review for the grammar or w/e leave it here. If you have a
different idea for what traitors should be leave it in the discourse
thread.

This is also my view on what traitors should strive to be as a gamemode
and the design decisions I think should be followed that would lead to
that. In terms of features I left it minimal mostly referencing things I
like and things I dislike about current traitors. There's not much
concrete here on "What should be added" so if you're here for that,
sorry to say but that's not what the doc is for.

---------

Co-authored-by: Princess Cheeseballs <66055347+Pronana@users.noreply.github.com>