Completely rewrite the pull request review guidelines (#558)
This pull request completely obliterates and rewrites the pull request
review guidelines from the ground-up.

Upstream has been tangled in endless and mindless bureaucracy for years.
The policy has quite simply failed us - game changes are caught up in
theoretical balancing and discussion instead of being disregarded or put
to the test.

The following policy ultimately aims to achieve the following:
- **Dramatically simplify policy.**
- The previous policy is a product of like 3 years worth of changes all
being merged into one amalgam of a document.
- **Reduces the required amount of maintainer input for small fry pull
requests, including content pull requests.**
- These types of pull requests are basically treated as a boolean MERGE
or CLOSE.
- **Removes the distinction between conceptual approval and code
approval for GitHub reviews.**
- They are rarely used and complicate policy. They have been rolled into
the exemptions category instead.
- **Reduces excessive voting.**
- Maintainers are encouraged to defer to other maintainers on things
they are unsure about. Discussions on gameplay changes are held
informally in organized spaces on Discord or the Forums. Maintainers are
heavily discouraged from voting unless absolutely necessary.
- **Encourages more activity on the Docs repo by enabling higher
throughput.**
- Maintainers are reminded that they can directly push to the repo for
most actions.
- **Shifts more work onto the Contributor.**
- We have been doing a lot of work basically fixing up a contributor's
pull requests. Pull requests often arrive with no justification or no
high-level overview of the changes so it's really hard to review it
because you have to figure out whatever they were thinking yourself.
- Contributors either provide justification for their pull request or it
gets closed.
- Contributors are expected to read and understand our codebase
conventions and guidelines to at least a 90% degree. If it is obvious
that they have not read them, the PR is to be closed instantly.